Refactor new for Date, Duration, Timestamp
